refactor : 도메인에서 사용하는 예외는 해당 도메인 패키지에서 관리
